Finalize changes for MVP (Reached 800 lines!)

+++ mainTermProject.py
@@ -12,31 +12,42 @@ from mainTermProjectArabicVer import *
 
 
 def distinguishSections(img):
+
     # perform OCR on screenshot of full page
+
     custom_config = r'--oem 3 --psm 1'
     string = pytesseract.image_to_string(img, config=custom_config)
+
+    # splits each section based on the new lines within the strings
 
     sectionsLst = string.split("\n\n")
     cleanSectionLst = []
 
     # create a new list of words separated by section
+
     for i in sectionsLst:
         cleanSectionLst.append(i.replace("\n", " ").split(" "))
 
-    # remove any unnecessary characters
+    # remove any unnecessary characters such as empty quotes
+
     for i in range(len(cleanSectionLst)):
         if " " in cleanSectionLst[i]:
             cleanSectionLst[i] = list(filter(" ".__ne__, cleanSectionLst[i]))
         if "" in cleanSectionLst[i]:
             cleanSectionLst[i] = list(filter("".__ne__, cleanSectionLst[i]))
 
+    # removes any sections with no words to prevent issues with
+
     if [] in cleanSectionLst:
         cleanSectionLst.remove([])
 
     # extra relevant data about the words on the image - coordinates
+
     dataDict = pytesseract.image_to_data(img, output_type=Output.DICT,
                                          config=custom_config)
     lstWords = list(dataDict["text"])
+
+    # helper function to help extract index of duplicated words
 
     def betterIdx(lst, word, num):
         # extracts the nth duplicate's index
@@ -51,10 +62,12 @@ def distinguishSections(img):
         return lastIdx
 
     # split each section with associated indices
+
     idxDict = {}
     dupWords = {}
 
     # links each section to the indices of the words in that section
+
     for i in cleanSectionLst:
         val = []
         for word in i:
@@ -65,6 +78,7 @@ def distinguishSections(img):
         idxDict[" ".join(i)] = val
 
     # extracts coordinates of each word in section
+
     coordDict = {}
     for key in idxDict:
         coordDict[key] = []
@@ -102,6 +116,7 @@ def distinguishSections(img):
     smallSections = []
 
     # find small sections based on section size
+
     for text in minMaxCoord:
         sectionSize = abs(minMaxCoord[text][2] - minMaxCoord[text][0]) * abs(
             minMaxCoord[text][3] - minMaxCoord[text][1])
@@ -112,6 +127,7 @@ def distinguishSections(img):
     newMinMaxCoord = {}
 
     # helper function to combine the texts in the sections and their coordinates
+
     def combineSections(minMaxDict, text1, text2):
         combineText = text1 + " " + text2
         combineCoord = [min(minMaxDict[text1][0], minMaxDict[text2][0]),
@@ -152,11 +168,17 @@ def distinguishSections(img):
 
 
 def mouseClickDetections():
+
     # in order to run the program, the user must press alt
+
     def on_press(key):
+
         # when alt key pressed start listening for a mouse press
+
         if key == keyboard.Key.alt:
+
             # moves mouse to the top left of the page
+
             mouse = Controller()
             mouse.position = (0, 1080)
             tts = gtts.gTTS("The mouse has been moved to the bottom left of the page", lang=userInterface.lang,
@@ -167,16 +189,20 @@ def mouseClickDetections():
 
             # takes a screenshot of the current page - region doesn't include address bar -
             # may need to adjust to fit screen
+
             img = pyautogui.screenshot('fullPgScreenshot.png', region=(0, 100, 1440, 900))
 
             # convert image from png to jpg so it can be used for the OCR - Grayscale is for better recognition
+
             pngToJpg = cv2.imread('fullPgScreenshot.png', cv2.IMREAD_GRAYSCALE)
             cv2.imwrite('fullPgScreenshot.jpg', pngToJpg)
 
             # extract sections of webpage
+
             on_press.stringWCoordDict = distinguishSections('fullPgScreenshot.jpg')
 
             # When the page is prepared for TTS
+
             tts = gtts.gTTS("You may now begin clicking", lang=userInterface.lang,
                             tld=userInterface.tld)
             tts.save("prepClicking.mp3")
@@ -184,9 +210,13 @@ def mouseClickDetections():
             os.remove("prepClicking.mp3")
 
             # page change variable, won't read any text until Alt is pressed again
+
             on_press.pgChange = False
 
         elif key == keyboard.Key.esc:
+
+            # message when escape is pressed
+
             tts = gtts.gTTS("Goodbye. Thank you for using the screen reader.", lang=userInterface.lang,
                             tld=userInterface.tld)
             tts.save("goodBye.mp3")
@@ -194,70 +224,101 @@ def mouseClickDetections():
             os.remove("goodBye.mp3")
             # remove extra files
             os.remove("fullPgScreenshot.jpg")
+
             # Stop mouse and keyboard listeners
             mListener.stop()
             return False
 
     def on_click(x, y, button, pressed):
+
+        # detects the mouse coordinates via the mouse listener
+
         mousePressCoord = [0, 0]
         try:
-            stringWCoordDict = on_press.stringWCoordDict
+            stringWCoordDict = on_press.stringWCoordDic
         except:
             stringWCoordDict = {}
 
+        # checks if the mouse has been pressed
+
         if pressed:
+
             # receive mouse press coordinates
+
             mousePressCoord = [x, y - 100]
 
-            # take a screenshot to compare with original screenshot to check if hyperlink was pressed
-            im = pyautogui.screenshot('comparePgScreenshot.png', region=(0, 100, 1440, 900))
+            # checks if page screen shot has been previously made to avoid any bugs and errors with clicking
 
-            # convert image from png to jpg so it can be used for the OCR - Grayscale is for better recognition
-            pngToJpg = cv2.imread('comparePgScreenshot.png', cv2.IMREAD_GRAYSCALE)
-            cv2.imwrite('comparePgScreenshot.jpg', pngToJpg)
+            if os.path.isfile('fullPgScreenshot.jpg'):
 
-            # check if two screenshots are the same - (abhi, StackOverFlow, 2020)
-            # https://stackoverflow.com/questions/1927660/compare-two-images-the-python-linux-way
-            with open('fullPgScreenshot.jpg', 'rb') as f:
-                content1 = f.read()
-            with open('comparePgScreenshot.jpg', 'rb') as f:
-                content2 = f.read()
-            if content1 != content2:
-                # perform text to speech on string
-                tts = gtts.gTTS("You have activated a hyperlink or have scrolled, press Alt to use the screen "
-                                "reader on the changed page", lang=userInterface.lang,
-                                tld=userInterface.tld)
-                tts.save("ttsPgChange.mp3")
-                playsound("ttsPgChange.mp3")
+                # take a screenshot to compare with original screenshot to check if hyperlink was pressed
 
-                # remove extra files created for screen reading process
-                os.remove('ttsPgChange.mp3')
-                on_press.pgChange = True
+                im = pyautogui.screenshot('comparePgScreenshot.png', region=(0, 100, 1440, 900))
 
-            # remove unnecessary files
-            os.remove('comparePgScreenshot.jpg')
-            os.remove('comparePgScreenshot.png')
+                # convert image from png to jpg so it can be used for the OCR - Grayscale is for better recognition
+
+                pngToJpg = cv2.imread('comparePgScreenshot.png', cv2.IMREAD_GRAYSCALE)
+                cv2.imwrite('comparePgScreenshot.jpg', pngToJpg)
+
+                # check if two screenshots are the same - (abhi, StackOverFlow, 2020)
+                # https://stackoverflow.com/questions/1927660/compare-two-images-the-python-linux-way
+
+                with open('fullPgScreenshot.jpg', 'rb') as f:
+                    content1 = f.read()
+                with open('comparePgScreenshot.jpg', 'rb') as f:
+                    content2 = f.read()
+
+                # if the pages differ then let the user know that they have activated a hyper link
+                # or scrolled
+
+                if content1 != content2:
+
+                    # perform text to speech on string
+
+                    tts = gtts.gTTS("You have activated a hyperlink or have scrolled, press Alt to use the screen "
+                                    "reader on the changed page", lang=userInterface.lang,
+                                    tld=userInterface.tld)
+                    tts.save("ttsPgChange.mp3")
+                    playsound("ttsPgChange.mp3")
+
+                    # remove extra files created for screen reading process
+
+                    os.remove('ttsPgChange.mp3')
+                    on_press.pgChange = True
+
+                # remove unnecessary files
+
+                os.remove('comparePgScreenshot.jpg')
+                os.remove('comparePgScreenshot.png')
 
         # check if page hasn't been changed with last click
+
         if not on_press.pgChange:
+
             # check if if the mousePress is within any of the predefined sections
+
             for string in stringWCoordDict:
                 if stringWCoordDict[string][0] <= mousePressCoord[0] <= stringWCoordDict[string][2] and \
                         stringWCoordDict[string][1] < mousePressCoord[1] < stringWCoordDict[string][3]:
+
                     # perform text to speech on extracted string
+
                     tts = gtts.gTTS(string, lang=userInterface.lang, tld=userInterface.tld)
                     tts.save("ttsOnString.mp3")
                     playsound("ttsOnString.mp3")
 
                     # remove extra files created for screen reading process
+
                     os.remove('ttsOnString.mp3')
+
             pgChange = False
 
     # used to listen for the key and mouse presses - (wowowo878787, StackOverFlow, 2019)
     # https://stackoverflow.com/questions/45973453/using-mouse-and-keyboard-listeners-together-in-python
+
     with keyboard.Listener(on_press=on_press) as kListener, mouse.Listener(on_click=on_click) as mListener:
         kListener.join()
-        mListener.join()
+        mListener.join
 
 
 def userInterface():
@@ -277,6 +338,8 @@ def userInterface():
     playsound("startSpeech.mp3")
     os.remove("startSpeech.mp3")
 
+    # instructions in english for the program
+
     instructions = """This application is a screen reader that helps you
                     navigate websites. In order to use this screen reader
                     go to the webpage that you want to be read and press alt.
@@ -286,6 +349,8 @@ def userInterface():
                     repeat the instructions in English press the spacebar. To repeat the instructions in arabic press A.
                     To leave the instructions press L."""
 
+    # instruction in arabic for the program
+
     instructionsAr = "هذا التطبيق عبارة" \
                      " عن قارئ" \
                      " شاشة يساعدك على تصفح مواقع الويب. لاستخدام قارئ الشاشة هذا ، انتقل إلى صفحة الويب التي تريد" \
@@ -294,6 +359,9 @@ def userInterface():
                      " الإنجليزية اضغط على مفتاح المسافة. لتكرار التعليمات بالعربية اضغط أ. لترك التعليمات اضغط على L."
 
     def on_press(key):
+
+        # deals with errors with keyboard listener
+
         try:
             # if D is pressed, accent can be
             if key.char == "d":
@@ -304,12 +372,16 @@ def userInterface():
                                                     accent you want to use for your screen reader, press enter and you can begin 
                                                     using the screen reader"""
 
+                # reads instructions in arabic
+
                 tts = gtts.gTTS(selectingAccentInstructions, lang=userInterface.lang,
                                 tld=userInterface.tld)
 
                 tts.save("selectingAccentInstructions.mp3")
                 playsound("selectingAccentInstructions.mp3")
                 os.remove("selectingAccentInstructions.mp3")
+
+            # if a is pressed read instructions in arabic
 
             elif key.char == "a":
 
@@ -324,7 +396,12 @@ def userInterface():
 
                 userInterface.screenReaderLanguage = "Ar"
 
+            # will leave the instructions page
+
             elif key.char == "l":
+
+                # message for leaving instruction page
+
                 tts = gtts.gTTS("Remember to press Alt when you are ready", lang=userInterface.lang,
                                 tld=userInterface.tld)
                 tts.save("leaveMessage.mp3")
@@ -334,7 +411,10 @@ def userInterface():
 
         except AttributeError:
 
+            # reading instructions in English, default language
+
             if key == keyboard.Key.space:
+
                 tts = gtts.gTTS(instructions, lang=userInterface.lang,
                                 tld=userInterface.tld)
                 tts.save("instructionsSpeech.mp3")
@@ -344,6 +424,8 @@ def userInterface():
                 # if the last spoken instruction was in English do OCR in English
 
                 userInterface.screenReaderLanguage = "Eng"
+
+            # an accent has been selected
 
             elif key == keyboard.Key.enter:
 
@@ -356,13 +438,19 @@ def userInterface():
                 os.remove("leaveMessage.mp3")
 
                 # when enter is clicked, accent is selected and keyboard listener stops
+
                 return False
+
+            # change option of accent
 
             elif key == keyboard.Key.up:
 
                 # options for the accent
+
                 accentOptions = ["American", "Australian", "British", "Indian", "French", "Spanish"]
                 userInterface.accent = accentOptions[userInterface.accentIdx % len(accentOptions)]
+
+                # if you want an American accent
 
                 if userInterface.accent == "American":
                     userInterface.lang = "en"
@@ -373,6 +461,8 @@ def userInterface():
                     playsound("accentChange.mp3")
                     os.remove("accentChange.mp3")
 
+                # if you want an Australian accent
+
                 elif userInterface.accent == "Australian":
                     userInterface.lang = "en"
                     userInterface.tld = "com.au"
@@ -381,6 +471,8 @@ def userInterface():
                     tts.save("accentChange.mp3")
                     playsound("accentChange.mp3")
                     os.remove("accentChange.mp3")
+
+                # if you want an British accent
 
                 elif userInterface.accent == "British":
                     userInterface.lang = "en"
@@ -391,6 +483,8 @@ def userInterface():
                     playsound("accentChange.mp3")
                     os.remove("accentChange.mp3")
 
+                # if you want an Indian accent
+
                 elif userInterface.accent == "Indian":
                     userInterface.lang = "en"
                     userInterface.tld = "co.in"
@@ -400,6 +494,8 @@ def userInterface():
                     playsound("accentChange.mp3")
                     os.remove("accentChange.mp3")
 
+                # if you want an Spanish accent
+
                 elif userInterface.accent == "Spanish":
                     userInterface.lang = "es"
                     userInterface.tld = "es"
@@ -408,6 +504,8 @@ def userInterface():
                     tts.save("accentChange.mp3")
                     playsound("accentChange.mp3")
                     os.remove("accentChange.mp3")
+
+                # if you want an French accent
 
                 elif userInterface.accent == "French":
                     userInterface.lang = "fr"
@@ -419,6 +517,7 @@ def userInterface():
                     os.remove("accentChange.mp3")
 
                 # selection changes everytime up is pressed
+
                 userInterface.accentIdx += 1
 
     with keyboard.Listener(on_press=on_press) as kListener:
